Added a solvable example board
This board is solvable and demonstrates the format used by default by
the parser. It is rated as being "extreme" by the source of the board.
The .board extension is not required but may be helpful in
distinguishing the files from code files.

On branch master
Changes to be committed:
	new file:   boards/extreme.board